fix(ast2700): correct pll rate calculation for system counter
plat_get_syscnt_freq2() programs cntfrq_el0 from the mpll rate, so a
wrong mpll value makes the kernel arch_timer (and every time-based
computation) run at the wrong frequency (observed 1010.06MHz instead of
1600MHz).

The root cause is the bitfield layout of union plat_pll_reg. Mixing
uint16_t and uint8_t base types makes the struct occupy 8 bytes instead
of 4, so the uint32_t 'w' overlay is no longer aligned with the n/p
fields and the N field is decoded incorrectly. For mpll (reg 0x16300:
m=768, n=11) the broken union read n=1, giving mult=768/2=384 instead of
768/12=64, which then overflowed the u32 rate computation to 1010065408.

Make all bitfield members uint32_t so the struct packs into a single
32-bit word (sizeof == 4) and the fields map to the correct register
bits.

+++ plat/aspeed/ast2700/plat_bl31_setup.c
@@ -127,13 +127,13 @@ entry_point_info_t *bl31_plat_get_next_image_ep_info(uint32_t type)
 union plat_pll_reg {
 	uint32_t w;
 	struct {
-		uint16_t m : 13;		/* bit[12:0]	*/
-		uint8_t n : 6;			/* bit[18:13]	*/
-		uint8_t p : 4;			/* bit[22:19]	*/
-		uint8_t off : 1;		/* bit[23]	*/
-		uint8_t bypass : 1;		/* bit[24]	*/
-		uint8_t reset : 1;		/* bit[25]	*/
-		uint8_t reserved : 6;		/* bit[31:26]	*/
+		uint32_t m : 13;		/* bit[12:0]	*/
+		uint32_t n : 6;			/* bit[18:13]	*/
+		uint32_t p : 4;			/* bit[22:19]	*/
+		uint32_t off : 1;		/* bit[23]	*/
+		uint32_t bypass : 1;		/* bit[24]	*/
+		uint32_t reset : 1;		/* bit[25]	*/
+		uint32_t reserved : 6;		/* bit[31:26]	*/
 	} b;
 };
